修复一些映射Bug
