+++ paconvert/utils.py
@@ -185,42 +185,3 @@ def log_debug(logger, msg, file=None, line=None):
     else:
         msg = "{}".format(msg)
     logger.debug(msg)
-
-
-def process_reduce_and_size_average(kwargs):
-    if "size_average" in kwargs:
-        size_average = kwargs.pop("size_average")
-        if "True" in size_average:
-            size_average = True
-        elif "False" in size_average:
-            size_average = False
-        else:
-            size_average = None
-    else:
-        size_average = None
-
-    if "reduce" in kwargs:
-        reduce = kwargs.pop("reduce")
-        if "True" in reduce:
-            reduce = True
-        elif "False" in reduce:
-            reduce = False
-        else:
-            reduce = None
-    else:
-        reduce = None
-
-    if size_average is not None or reduce is not None:
-        if size_average is None:
-            size_average = True
-        if reduce is None:
-            reduce = True
-
-        if size_average and reduce:
-            reduction = '"""mean"""'
-        elif reduce:
-            reduction = '"""sum"""'
-        else:
-            reduction = '"""none"""'
-
-        kwargs["reduction"] = reduction
